src/form-recognizer-scratch/requirements.txt
azure-ai-formrecognizer